clase5/app/requirements.txt
Flask==2.1.1
Flask-swagger-ui
pymysql==1.0.2
werkzeug==2.1.2
mysql-connector-python==8.0.32
Flask-restx==0.5.1